chore(zero): move `server-schema` to `zero-schema` package
The server schema needs to be available to zero-cache and, in the future, other packages for different backends (pg vs mysql vs ...).
