+++ packages/z2s/src/sql.ts
@@ -7,7 +7,7 @@ import baseSql, {SQLItemType} from '@databases/sql';
 import {assert, unreachable} from '../../shared/src/asserts.ts';
 import {isPgNumberType, isPgStringType} from '../../zero-cache/src/types/pg.ts';
 import type {LiteralValue} from '../../zero-protocol/src/ast.ts';
-import type {ServerColumnSchema} from './schema.ts';
+import type {ServerColumnSchema} from '../../zero-schema/src/server-schema.ts';
 
 export const Z2S_COLLATION = 'ucs_basic';
