+++ packages/zero-pg/src/schema.ts
@@ -1,7 +1,7 @@
 import {assert} from '../../shared/src/asserts.ts';
 import type {Enum} from '../../shared/src/enum.ts';
 import {must} from '../../shared/src/must.ts';
-import type {ServerSchema} from '../../z2s/src/schema.ts';
+import type {ServerSchema} from '../../zero-schema/src/server-schema.ts';
 import {formatPg, sql} from '../../z2s/src/sql.ts';
 import * as PostgresTypeClass from '../../zero-cache/src/db/postgres-type-class-enum.ts';
 import {dataTypeToZqlValueType} from '../../zero-cache/src/types/pg.ts';
